Fix test using ? glob with ObjectToURI
Passing a ? wildcard through ObjectToURI does not work because passing
the constructed URI through urlparse.urlparse() interprets ? as a query
and strips it off since there is no query string following it. The *
glob which is actually used in the very next call doesn't have that
problem so just use that instead.