Only use proxy settings from boto if they are valid.
Do not assume .boto always has proper proxy settings. For example, this
is not the case for Chromium+depot_tools users, which are not even
required to have a boto configuration file around.

If the ProxyInfo object created by using the settings from boto is not
valid, do not pass |proxy_info| to httplib2.Http and instead rely on it
figuring out the proxy settings on its own (at present, it will look at
the user's proxy environment variables).

BUG=chromium:443523