Perform only a single listing to check for bucket subdirectories
This change performs a prefixed listing of both prefixes and
object names in order to reduce subdirectory folder detection to a
single HTTP call.

This change improves performance and reduces cost with two tradeoffs.
First, detection of a bucket subdirectory based on the presence of a
_$folder$ object is now eventually consistent. Second, copying an
object to a destination URL that is not a bucket subdirectory but
is a prefix of many existing objects now requires iterating through
the listing of all objects matching the prefix to ensure that no
_$folder$ object exists.  This could be slow if there are many matching
objects.