Fix rm -r with missing URLs; fix shared_attrs in command._SequentialApply
This change corrects the behavior of rm -r when run in multi-threaded
mode or with the continue_on_error (-f) flag. Previously, rm would
attempt to delete a bucket, even if listing that bucket returned a
404.

It also causes the multi-threaded or force-continue rm to correctly
note when some files could not be removed.

It also adds a counter of the number of objects not successfully
removed to be output in the case of an error.

Finally, it fixes a bug with shared attributes in command.SequentialApply,
where if the last argument resulted in an exception that updated a
shared attribute, the update would not be committed.