Use temporary files for all downloads
All downloads write to a temporary file while the download is in progress. A
file at a download's permanent file location is deleted before the download
starts, so that extra disk space is not required in the case of a preexisting
file at the permanent location.

Logging still uses the download's permanent file location.