Use multiple threads even when multiprocessing is unavailable.
This change fixes a bug where multiple threads would not be used
unless multiprocessing was available.

It does so by creating thread-safe, process-unsafe globals in the
parallelism framework for the case where only a single process
is used. These globals act like the multiprocessing.Manager
globals in InitializeMultiprocessingVariables, and are available
whenever multiprocessing is unavailable, regardless of whether
multiple threads are actually used by Apply.