Implement slice, fan, and both parallel strategies in perfdiag.
This change implements the slice, fan, and both strategies in perfdiag, which
can be used to modify how throughput tests use threads/processes. Two new
command line options, -p and -y, specify the parallelism strategy and the
number of slices to use, respectively.

The documentation for the -n command line option has also been changed, now
documenting it as the number of objects to use rather than the number of
iterations to run. The underlying behavior hasn't changed.

This also fixes a bug where perfdiag stats were incorrect when parallelism was
used, and modifies perfdiag cleanup to be done in parallel if parallelism is
specified.