Fix perfdiag rthru_file -p slice. Upload test files only when needed.
This change fixes a performance bug that occurred because the
rthru_file test overwrote the original file on disk, as opposed to
recreating it.

It also eliminates an unnecessary upload when running both rthru_file
and wthru_file tests. The objects created by the wthru_file test are
used for the rthru_file download.