Enabled sliced downloads by default, rename them, and add documentation
This change enables sliced downloads by default, with a default 4 slices.

It globally renames parallel object download to sliced object download
for clarity.  It also documents sliced object downloads.

It also adds a suggestion to install crcmod when it is not available and
sliced downloads would be useful.

Finally, it fixes a bug where suggestions to use sliced downloads or
parallel composite uploads would be output once per process as opposed
to once globally.