Fix gsutil interactions with transcoding and compression
This change sends accept-encoding: gzip to the service only
when the requested object has content-encoding: gzip or the
entire range of the object is being requested. This
prevents compressive transcoding by the service.
Compressive transcoding can cause responses to range requests
to instead issue all of the bytes of the object, which,
while legal according to the HTTP spec, breaks resumability.

The change also adds cache-control no-transform to objects
that are uploaded (via the -z or -Z options) with gzipped
content-encoding. This ensures that objects that are stored
doubly-compressed are served as-is, without removing the first
layer of compression. This is necessary because removing
the first layer of compression would cause the content served
to the user to differ from the cloud-stored hashes for the
doubly-compressed object.

Fixes https://github.com/GoogleCloudPlatform/gsutil/issues/324