Reduce number of threads on Linux systems by 2x (from 10 to 5)
Since a long-standing process starvation bug was fixed in
gsutil 4.14, the number of active threads on Linux systems
increased from 10 to 10*num_CPUs. Lock contention and
thread-spawning failures increased because a much higher number
of threads are being used to do work. This change reduces the
default number in an effort to reduce those errors while still
maintaining good performance for large workloads.