Move progress prints to a central UI thread
This change establishes a global queue that is responsible
for printing progress messages, as opposed to having worker threads
printing messages directly.

This change should have no visible impact now, but lays the
groundwork for allowing worker threads (in one or more processes)
to pass richer status messages. These messages could then be
aggregrated and displayed to the user or sent to an analytics pipeline.