Remove an HTTP metadata get call from cp, acl, setmeta
This changes the name expansion iterator to accept fields
to enumerate up-front so that a later metadata get call can
be avoided. The cp, acl, and setmeta commands will make one
less get call per object at the expense of listing more
fields for objects that may be filtered via a wildcard.