Fix bug in rsync -d iteration
This change fixes a bug where the first run of gsutil rsync -d
could erroneously remove the final object in the destination listing.
Subsequent executions of the same rsync -d command would remedy the
issue.

In order for this to happen, two conditions had to be met:
1) Rsync reached a point where it had iterated over N-1 destination
objects before iterating over all source objects
2) The previous object comparison was of two objects with different
filenames

At this point, rsync mistakenly considered itself to be out of
destination objects and removed the last destination object regardless
of whether or not it was also present in the source listing.