Add task estimation capability for large operations
This change adds an optional estimation thread to commands using
multiprocessing and/or multithreading. After a threshold of tasks
are iterated by the main producer thread, it spawns the estimation
thread. This thread performs listing calls to estimate the total
amount of work that the call to Command.Apply will perform.
It writes a one-line message to the console upon
completing this estimation. In the future, it can be used to
estimate the total progress and time remaining for the command.

The threshold is configurable and the esitmation thread will never
spawn if the threshold is set to 0. By default, the threshold is set
near the number of tasks that the producer thread can iterate
before the global tasks queue becomes completely full.