No longer rely on pbr in mock fork.
This change updates the mock submodule to include changes that make it
no longer rely on pbr - this prevents inconsistent behavior that might
result when not using a pip-installed version (such as `gsutil test`
failing to run). Pip installations of gsutil will still use the canonical
mock and pbr packages, which will be fine, since or submodule variations
only vary enough to cover problems deriving from lack of pip installation.