Implementation of file modification time with the rsync command.
This changes the default change detection algorithm from file size to
file modification time (mtime). If mtime is available at the source and
destination, it will be used; otherwise checksums will be computed and
compared when possible. If checksums are unusable, file size will be
used as a last resort.

Using mtime will make synchronizations both fast and accurate. For files
not currently synced with their mtime, a checksum will be used for the
first comparison, and the destination will be updated with the mtime for
future comparisons.