Temporarily relax test_cp_v_option
This relaxes the ordering constraints for the Created message.
Since the UIThread now manages progress for sliced object
downloads, it's possible a progress message will print after
the Created message.

To fix, we'd need to ensure the Created message is also
serialized through the global status queue and manage ordering
in the UIThread. More likely, we'll refactor progress prints
in general.