Implement POSIX attribute preservation.
This change will add the -P flag to the cp, mv, and rsync commands. The
attributes access time (atime), group ID (gid), user ID (uid), and mode
will be saved to custom metadata attributes for files uploaded to cloud
objects. Upon download these attributes will be validated so that access
to the file will not be orphaned at any point. For rsync, if any files
would be orphaned this error appears before any actual sync. Whereas
with cp and mv, errors are displayed as they are detected. Any
validation error will be fatal and terminate the command.

Attributes will only be uploaded and downloaded from *nix machines as
Windows does not fully support the POSIX standard. Thus, only atime and
mtime can be saved when syncing with a Windows machine.