Refactor single process multi-threaded parallelism
This change migrates parallelism code in the single
process case to use a similar approach as the multi-process
case. Instead of creating a new task queue for each call
to Apply, a single task queue per recursive apply level
is created. One worker thread pool is spawned per
recursive apply level and that pool is responsible for
performing all work at that level.

This fixes a bug where a a new multiprocessing.Queue was
created for each call to Apply in the single process case,
and added to the global queue list so it was not eligible
for garbage collection.