Convert test object creation to idempotent model to reduce flakiness
This change addresses test flakiness that can occur when a test
object is created and, due to a network timeout or distributed systems
failure, the creation is retried even though the original request
succeeded.

To address this, this change adds a generation precondition parameter
to object creation in all gs integration tests.  The default value of
the parameter is 0, which creates a new object only if there is
no existing object with that name.  On precondition failure
the tests assume the live object was correctly created by a previous
attempt.

Starting with this change, tests which overwrite a test object with
the same name must supply a precondition of the original object's
generation.

A consequence of this change is that name collision from concurrently
running tests (for example, two users running tests in the same
project) will produce undefined behavior.  This is mitigated by the
test framework's randomized bucket and object naming.