Fix #431 - allow unicode in custom metadata from top-level -h option.
We already allow these characters in custom metadata in the setmeta
command; the -h flag to specify header values should allow the same set
of characters.

This change adds the HandleHeaderCoding method, which checks for proper
encoding of headers. This also updates the HandleArgCoding method, to
which we stopped passing headers at some point. It also previously
checked for ASCII-only characters when evaluating the -p flag -- this
was intended to apply to project IDs, but we now have other commands with
-p flags that represent things other than project IDs. I've moved the
project ID ASCII-only checks to the commands which have -p as an option
that refers to project ID.