Use correct provider in CreateObject in unit tests
This was causing a tab completion test against s3 to fail, as it was
expecting an s3://... value and instead was using a gs://... value.

This change also uncovered a test failure that showed we were requesting
all object metadata when retrieving it for the "cat" command. We now
only request the necessary subset of metadata fields when running "cat".