Fix cp storage class changes introduced w/ 8896457
Because Boto is now aware of the storage class header for GCS's XML API
(as of commit 8896457), it started utilizing its default behavior of
specifying STANDARD as the default storage class header value for object
copy requests. In contrast, we want the bucket-default storage class to
be used (no storage class is supplied in the request; it's determined by
the server). This fixes a bug where STANDARD was being used as the
storage class for all copies where no storage class was explicitly
specified (via the -s flag for cp). Note that this bug is not present in
any released version of gsutil; it was introduced after releasing 4.26,
and we have not released 4.27 as of this commit.