Print helpful error for improperly formatted CORS.
Fixes #441. When running "gsutil cors set", using a CORS configuration
containing a JSON object not wrapped in a list would throw an ambiguous
error: 'unicode' object has no attribute 'iteritems'. This adds some
validation to make sure the parsed JSON is a list, and if not, we throw
a helpful message.