Add transport compression support. (#486)
* Add transport compression support.

Included:
- Integration with cp, rsync, and perfdiag commands.
- Resource limiting for parallel, transport compressed uploads.
Transport compressed uploads buffer data into memory to serve in a
chunked manor. For parallel uploads, many chunks may exist in memory
concurrently. To prevent excessive memory consumption, a lock is used
to restrict the number of transport compressed uploads that may be ran
in parallel.
- Dynamic progress printing for transport compressed uploads. To more
accurately represent the progress of a transport compressed upload, the
progress printing is scaled to display the effective upload rate.

* Add annotations to skip transport compression tests when using the XML
api.