Only add bundled crcmod to path when necessary.
Addresses issue #462.

As the bundled crcmod module in third_party doesn't contain its compiled
C extension, we should only import it when no other crcmod module is
available on the user's system.

Further, the previous logic was misleading -- it prepended the bundled
module's path to sys.path; if the system already had a crcmod module and it
was successfully loaded, adding the bundled crcmod's path to the front
of sys.path wouldn't cause the system crcmod module to be unloaded in
favor of using the bundled one.