Corrects info log messages for computing checksums (#551)
Corrects info log messages for computing checksums

When computing the MD5 or CRC32C of a file, we use `logger.info()` to tell the user what's going on. The log message is mixed up on a couple of these messages, indicating MD5 when we're actually computing CRC32C, or vice versa.  The log messages also mistakenly mixed up src and dst in a couple of places, printing out the wrong file/object name.