Adds local patch for boto/boto#3831.
In working on Python 3 support for gsutil, we found that the order of
AuthHandler subclasses is not represented consistently by the
`__subclasses__` method across all Python 3 versions, leading to the bug
mentioned above. We need to ensure that OAuth2 credentials should be
chosen for use when both OAuth2 and HMAC credentials for GCS are present
in the supplied Boto config file(s). Without this, it would be possible
for two consecutive gsutil invocations to use different auth handlers
without any code or configuration changes in Python 3.4 or 3.5.

Note that this did not affect any existing users because Python 3
support is not yet fully implemented.