Fixes #575.
In certain scenarios, pwd.getpwuid() may be unable to find the user's
UID in the system user database, causing that method or any method that
uses it (e.g. getpass.getuser()) to raise a KeyError.

Also, we should be checking for `os.geteuid() == 0` to test if the user
is root on non-Windows systems.