Don't clobber local files with differing size
As discussed in #591, `gsutil cp -n` will overwrite an existing local
file if its size differs from the size of the source file.  This
behavior is contrary to user expectations, POSIX convention, and the
documentation.

Remove the special case for files with differing size, which was added
in a768a58 to support resuming partial downloads.  This code has not
been useful since 3e3d6dc, which started using temporary files (named
with _.gstmp) for all downloads.

Fixes: #591

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>