comparison test fixed
binary file was read and compared to string in assertion. Python 2 str
is binary based so it would pass, Python 3 str is unicode based so
the assertion failed since it was comparing unicode to binary. Added
specifier to comparison string creation to ensure string is created as
binary string in both Python 2 and 3.