Fixes #606 - use SNI for boto cxns if possible
Boto https connections weren't using SNI before. With the recent-ish
introduction of OpenSSL 1.1.1, we start to see SSL errors when making
requests to storage.googleapis.com (GCS's XML API) where the server
hostname was not specified. This pulls in
https://github.com/boto/boto/pull/3843 as a hotfix into gsutil, since
there isn't a released version of Boto where this PR is included yet,
and we'd like to have it sooner rather than later.