Getting Py-Six-Current up to date - PR05 (#637)
* Defining cmd in a more concise manner

* Wrapping `cmd` list elements in six.ensure_text

This is to prevent issues with unicode in the command text when
running in Python 2.7.

* Allowing unicode or binary text type as "contents"

Added automatic conversion to binary type if contents is either binary
or text type. Any other types will still raise TypeError.

* encoding to ascii at string.printable caused error in Python 3

encoding ascii and then requesting character at a location in the string
was causing a return value of an integer instead of the character at
the desired location. Encoding the result before assigning to contents
produces the desired result in Python 2 & 3.

* Changed order of temp file creation for macOS

Temp file creation only creates 2 files on macOS because of the way the
os handles certain unicode characters. Comparison was failing because
the first file was not in the comparison set.

* comparison test fixed

binary file was read and compared to string in assertion. Python 2 str
is binary based so it would pass, Python 3 str is unicode based so
the assertion failed since it was comparing unicode to binary. Added
specifier to comparison string creation to ensure string is created as
binary string in both Python 2 and 3.

* Verify type of list elements in CreateTempFile

file_name can be a tuple or other iterable and no checks were around
the individual elements. Adding a check to ensure that elements are
str before join command.