Fix failure flag (#640)
* sequential_success was being overridden

sequential_success is the flag that determines if the program prints
success or failure at the end of the run. It also determines the exit
code overall. Since it was being set to an empty array, it counted as
a falsy and caused the failure message to print and the exit code of 1.

* Ignoring .DS_Store file in macOS during update check