Catch up CI branch with master (#665)
* Explicit wording in rm docs about deleting buckets (#647)

* Fix docs formatting to convert to better HTML. (#654)

These changes seem to make Sphinx happy, and I'm tired of dealing with
all the manual fixes to generated HTML docs.

* Additional checks around ShouldProhibitMultiprocessing() (#656)

User reporting that this check crashes in FreeDesktop:
https://github.com/GoogleCloudPlatform/gsutil/issues/655
Cause: this check assumes that the first line is the correct line with
'NAME="<linux ver>"'. In most cases, this is correct, however this
assumption cannot be relied upon. Additional checks around the contents
of each line were added to look for 'NAME='.

Generic exception added to catch any other issues from this method in
the future (with exception information added to logging).

* Updated Kokoro build script paths. (#659)

Build paths all have a typo, per b/127678469#comment7. This is
preventing Kokoro builds.

* Added continuous build configs (#663)

* Added continuous build configs

Added Kokoro CI build configs to trigger every time a push happens
on our repository. Although this will largely overlap with the
presubmit tests running on every PR, it may help us in the case of
code being pushed directly to the repo without a PR.

* Adjusted CI test timeouts.