Move vendored directory into gslib. (#666)
With our last couple of CLs to start vendoring boto, we overlooked that
our setup.py was excluding packages under the top-level third_party
directory.  This moves the vendored libraries to gslib/vendored so that
they're always included with the gslib code, even for pip installs.

This also cleans up some of the logic around having to guess whether we
were using boto that came with a pip installation or a packaged
installation.