Update storage_v1_* apitools generated modules.
This change updates the JSON API client files in
gslib/third_party/storage_apitools to be consistent with the source
documents (via the JSON API discovery doc) and updates docstrings to
the new style that apitools generates (using raw strings), which makes
this change a bit noisy.

This commit also adds the "includeTrailingDelimiter" field to some
message types, but that is uninteresting because it's unused in the
code.