Add notes to docs mentioning `iam ch` limitation.
The `iam ch` command does not work with resources that have conditions
in their IAM policy bindings, and will throw an error if users try to do
so.

Also fixes the helptext to surround commands in double-backticks instead
of double-quotes so that they will be rendered correctly when generating
the web docs.