Revised get_random_ascii_characters function (#718)
* Revised get_random_ascii_characters function

get_random_ascii_characters actually needed to return a string
representation of a list of characters, rather than an actual
list of characters for our testing purposes. The logic has been
revised to represent this need, bearing closer to is orignal
authorship with some revisions for Python 3 compatibility.

* Remove unused charset variable, edit docstring

Variable charset was assigned but never used.

Docstring header specified string, but not if it was binary or text.