Introduce gzip_compression_level config option
This allows configuring the GZIP compression level used if
compression is enabled. The Python stdlib default is 9, the
highest possible; this makes sense if someone is entirely
optimizing for storage space cost.

For speed, reducing compression level to 6, which is the
default used by the gzip tool, improves compression speed
substantially for many files without losing much space.