Added explicit "bash" to config_gen.sh call (#753)
Per b/131015356 , currently when the kbuilder user on Kokoro tries
to call config_generator.sh directly, it gets permission denied.
This seems to be due to kbuilder not having execute permission on
this file.

We can work around this by execing bash with the script path as an
argument. Additionally, I added execute permission to both test scripts.