Removed flakey assert (#764)
For b/131325628. We noticed in the CI pipeline on Windows that the test
test_ui_resumable_download_break_with_m_flag tended to flake fairly
frequently due to a race condition where the artifical download halt
signal was sent before the "0% Done" text was printed. Discussing with
Matt we agreed this assert didn't add very much value, and was fine to
simply remove.