Fixed integ test for 'cp -L' flag (#774)
For b/130055461. While doing Python 2/3 compat changes, we discovered
that the 'cp -L' flag had integration tests, but they failed to catch when headers
were wrapped in b' ' artifacts. This was because the asserts were only checking
a partial string slice. We fixed this by matching against the full source and dest.