Remove test dir from test_update (#789)
The presence of many new scripts in the test directory upset and confused our pre-release tests. In this PR we're asking it not to look in this folder since it's all CI stuff that is tested separately anyways.

There was also a small bug with an earlier PR that introduced a regression for signing URLs in Python 2.x. Added a fix for this as well.