Add debugging mechanisms and fix tabcomplete tests.
We started running tabcomplete tests by specifying the interpreter used
to invoke gsutil; this ended up causing argcomplete to provide no
matches, as it was trying to find a Completer for something like "python
gsutil ls" (which didn't exist) rather than "gsutil ls" (which existed).

This is the 4th time I've had to do "clever" things to see what
argcomplete was printing out for me in order to debug test failures, so
I added in a boolean in the test code to toggle the add'l debug prints.