Prefix invalid tabcomplete bucket name
Per PR 766's description, some XML tests fail when over 1000 buckets are
present. This is one of the tests that does this, and we hoist the
bucket name close to the top of the list py prefixing it with "aaa-".
Also fixed MakeTempName to always accommodate enough room for "aaa-".