Only try calling os.setsid if that function exists
It does not exist on Windows, so the integration tests were failing.
Not sure why Kokoro was reporting successes; maybe it was only looking
up the status of the first tests to finish (which are usually the Linux
tests)?  Regardless, this fixes the issue of trying to call os.setsid on
Windows.