Add .pylintrc to check a subset of warnings. (#858)
Also adds a shell script that selects which files should be passed to
the linter.

TODOs:

We still need to gradually go through the rest of the warnings that are
commented out and see which ones can safely be uncommented, which ones
are revealing actual issues, and which ones are just noise (and thus
should stay disabled) due to some of the odd imports and monkeypatching we do.

We should also run this script as part of the TravisCI presubmits...
probably under both Python 2 and 3? I haven't tried running it under Py3
yet.