Fix incorrect rendered number of objects when using perfdiag -i
Because the value passed to -n wasn't stored in the output file (when
-o is provided), running `perfdiag -i` with that file would
render inconsistent data for throughput tests.

Like in the example below, where 5 is default value for -n.

```
Copied 5 20 MiB file(s) for a total transfer size of 1.95 GiB.
```

For the record, when provided both -n and -i, perfdiag would render
something like in the example below, where 999 is value passed to -n.

```
Copied 999 20 MiB file(s) for a total transfer size of 1.95 GiB.
```

Now the number passed to -n is stored in the results, and used to
render the proper output. The previous behavior is preserved for
backward compatibility reasons, in case of running perfdiag -i with an
old file.