Specify the purpose of -p option for -D example
The examples for -D option is confusing, as -p option is also specified. 

Some users end up thinking that these two options should be used together. This is not always the case (e.g. for bucket with uniform bucket-level access),  and an error will be returned in case OWNER permission is missing as OWNER permission is required for preserving ACLs. 
This cl should clarify better the examples and avoid further support tickets.