gslib: boto_util: implement a HasUserSpecifiedGsHost() helper (#1265)
The aim is to avoid having to include 'HAS_NON_DEFAULT_GS_HOST' from the
'gslib.tests.util' package. In general, test-code should not be included in
the main library/program code (if/where possible).

In this case, the logic for 'HAS_NON_DEFAULT_GS_HOST' is that of a
'HasUserSpecifiedGsHost()' helper, where the
MaybeCheckForAndOfferSoftwareUpdate() method is checking whether to
interact with the user. The comment also mentions:
```
    # Don't try to interact with user if:
    # - gsutil is not connected to a tty (e.g., if being run from cron);
    # - user is running gsutil -q
    # - user is running the config command (which could otherwise attempt to
    #   check for an update for a user running behind a proxy, who has not yet
    #   configured gsutil to go through the proxy; for such users we need the
    #   first connection attempt to be made by the gsutil config command).
    # - user is running the version command (which gets run when using
    #   gsutil -D, which would prevent users with proxy config problems from
    #   sending us gsutil -D output).
    # - user is running the update command (which could otherwise cause an
    #   additional note that an update is available when user is already trying
    #   to perform an update);
[1] # - user specified gs_host (which could be a non-production different
    #   service instance, in which case credentials won't work for checking
    #   gsutil tarball).
    # - user is using a Cloud SDK install (which should only be updated via
    #   gcloud components update)
```

This change addresses condition mentioned in [1].

Trying to include 'gslib.tests.util' also pulls-in some various other
modules like 'mock_storage_service', and even creates a 'mock_connection'
singleton, which aren't typically needed when running gslib in the main
library.

The 'gslib.boto_util' module seems like a good place to put the logic, to
be shared in the 'gslib.tests.util' and 'gslib.command_runner' modules.

Signed-off-by: Alexandru Ardelean <aardelean@deviqon.com>