Fix formatting issue in generated docs (#1526)
The linter command does not format correctly on cloud.google.com, and while PR #1281 attempted to fix this, it appears it did not. My guess is the indentation issue is related to the entire item being indented an extra space (owing to it being a 2-digit number). To circumvent this, I'm proposing we combine two separate-but-related items above into a single item.