Ensured a named pipe has a size greater than 0
Before the change, the if statment did not check that a named pipe has a
size greater than 0, causing it to exit before moving any data and
causing the write thread to be stuck.

After the change, there is a check that the named pipe has a size
greater than 0, avoiding the issue described above.