Updated the cat -r command error behavior (#1531)
* fixed -r bugs

* fixed based off comments

* Edited gsutil cat -r - -h behavior and fixed timout issues

Before the change, running cat -r - -h would not print the
header of the object. Moreover, some tests would time out.

After the change, running cat -r - -h prints the header of
the object and the tests do not timeout.

Example:
If gs://folder/object1.txt's content is 0123456789,
and the user runs gsutil cat -r - -h, then the output
will now be:
==>gs://folder/object1.txt<==
0123456789

* Fixed code to pass kokoro tests

cat_object is a named pipe. Therefore, the size attribute
could be None. I added a check to make sure it is not None
before making a comparison with it.

* Edited the if statement to pass kokoro tests

cat_object might not have a size attribute. Therefore,
the new if statement checks if the cat_object has
a size attribute. If it does not, then it says it to None.

* Ensured a named pipe has a size greater than 0

Before the change, the if statment did not check that a named pipe has a
size greater than 0, causing it to exit before moving any data and
causing the write thread to be stuck.

After the change, there is a check that the named pipe has a size
greater than 0, avoiding the issue described above.

* Addressed feedback and changed if statement

* Replaced "float('inf')" with -1 in the if statement.